Move SantoDomingoDatetime to database.py
